refactor: organize management comp/svc directives
