+++ src/app/management/directives/data-manager/data-manager.service.directive.ts
