+++ src/app/management/directives/transactional-data-manager/transactional-data-manager.component.directive.ts
@@ -5,13 +5,13 @@
  * https://opensource.org/licenses/MIT
  */
 
-import { Directive, OnInit } from '@angular/core';
-import { Observable, of } from 'rxjs';
-import { map, tap } from 'rxjs/operators';
+import { Directive } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
-import { EntityFormDialogComponent } from '../../shared/dialogs/entity-form/entity-form-dialog.component';
-import { EntityFormDialogConfig } from '../../shared/dialogs/entity-form/EntityFormDialogConfig';
-import { DataManagerComponentDirective } from './data-manager.component-directive';
+import { Observable } from 'rxjs';
+import { tap } from 'rxjs/operators';
+import { EntityFormDialogComponent } from '../../../shared/dialogs/entity-form/entity-form-dialog.component';
+import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form/EntityFormDialogConfig';
+import { DataManagerComponentDirective } from '../data-manager/data-manager.component.directive';
 
 /**
  * Base for data manager template components.
