+++ src/app/management/routes/customers/management-customers.component.ts
@@ -8,7 +8,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Customer } from 'src/models/entities/Customer';
-import { DataManagerComponentDirective } from '../../directives/data-manager.component-directive';
+import { DataManagerComponentDirective } from '../../directives/data-manager/data-manager.component.directive';
 import { ManagementCustomersService } from './management-customers.service';
 
 @Component({
