+++ src/app/management/routes/customers/management-customers.service.ts
@@ -9,7 +9,7 @@ import { Inject, Injectable } from '@angular/core';
 import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-service-injection-tokens';
 import { IEntityDataApiService } from 'src/app/api/entity.data-api.iservice';
 import { Customer } from 'src/models/entities/Customer';
-import { DataManagerServiceDirective } from '../../directives/data-manager.service-directive';
+import { DataManagerServiceDirective } from '../../directives/data-manager/data-manager.service.directive';
 
 @Injectable()
 export class ManagementCustomersService
