+++ src/app/management/routes/images/management-images.component.ts
@@ -9,14 +9,14 @@ import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute } from '@angular/router';
-import { catchError, map, tap } from 'rxjs/operators';
-import { Image } from 'src/models/entities/Image';
-import { ImageFormComponent } from 'src/app/shared/components/image-form/image-form.component';
-import { COMMON_DISMISS_BUTTON_LABEL, COMMON_ERROR_MESSAGE, COMMON_WARNING_MESSAGE } from 'src/text/messages';
-import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form/EntityFormDialogConfig';
-import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager.component-directive';
-import { ManagementImagesService } from './management-images.service';
 import { of } from 'rxjs';
+import { catchError, tap } from 'rxjs/operators';
+import { ImageFormComponent } from 'src/app/shared/components/image-form/image-form.component';
+import { Image } from 'src/models/entities/Image';
+import { COMMON_DISMISS_BUTTON_LABEL, COMMON_ERROR_MESSAGE } from 'src/text/messages';
+import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form/EntityFormDialogConfig';
+import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager/transactional-data-manager.component.directive';
+import { ManagementImagesService } from './management-images.service';
 
 @Component({
   selector: 'app-management-images',
