+++ src/app/management/routes/product-categories/management-product-categories.component.ts
@@ -11,7 +11,7 @@ import { ActivatedRoute } from '@angular/router';
 import { ProductCategoryFormComponent } from 'src/app/shared/components/product-category-form/product-category-form.component';
 import { EntityFormDialogConfig } from 'src/app/shared/dialogs/entity-form/EntityFormDialogConfig';
 import { ProductCategory } from 'src/models/entities/ProductCategory';
-import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager.component-directive';
+import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager/transactional-data-manager.component.directive';
 import { ManagementProductCategoriesService } from './management-product-categories.service';
 
 @Component({
