+++ src/app/management/routes/product-lists/management-product-lists.component.ts
@@ -13,11 +13,11 @@ import { of } from 'rxjs';
 import { catchError, map, tap } from 'rxjs/operators';
 import { ProductListFormComponent } from 'src/app/shared/components/product-list-form/product-list-form.component';
 import { ProductList } from 'src/models/entities/ProductList';
-import { COMMON_DISMISS_BUTTON_LABEL, COMMON_ERROR_MESSAGE, COMMON_WARNING_MESSAGE } from 'src/text/messages';
+import { COMMON_DISMISS_BUTTON_LABEL, COMMON_ERROR_MESSAGE } from 'src/text/messages';
 import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form/EntityFormDialogConfig';
 import { ProductListContentsDialogComponent } from '../../dialogs/product-list-contents/product-list-contents-dialog.component';
 import { ProductListContentsDialogData } from '../../dialogs/product-list-contents/ProductListContentsDialogData';
-import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager.component-directive';
+import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager/transactional-data-manager.component.directive';
 import { ManagementProductListsService } from './management-product-lists.service';
 
 @Component({
