+++ src/app/management/routes/products/management-products.component.ts
@@ -9,14 +9,14 @@ import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute } from '@angular/router';
-import { catchError, map, tap } from 'rxjs/operators';
-import { Product } from 'src/models/entities/Product';
-import { ProductFormComponent } from 'src/app/shared/components/product-form/product-form.component';
-import { COMMON_WARNING_MESSAGE, COMMON_DISMISS_BUTTON_LABEL, COMMON_ERROR_MESSAGE } from 'src/text/messages';
-import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form/EntityFormDialogConfig';
-import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager.component-directive';
-import { ManagementProductsService } from './management-products.service';
 import { of } from 'rxjs';
+import { catchError, map, tap } from 'rxjs/operators';
+import { ProductFormComponent } from 'src/app/shared/components/product-form/product-form.component';
+import { Product } from 'src/models/entities/Product';
+import { COMMON_DISMISS_BUTTON_LABEL, COMMON_ERROR_MESSAGE } from 'src/text/messages';
+import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form/EntityFormDialogConfig';
+import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager/transactional-data-manager.component.directive';
+import { ManagementProductsService } from './management-products.service';
 
 @Component({
   selector: 'app-management-products',
