+++ src/app/management/routes/salespeople/management-salespeople.component.ts
@@ -9,14 +9,14 @@ import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute } from '@angular/router';
-import { catchError, map, tap } from 'rxjs/operators';
-import { Salesperson } from 'src/models/entities/Salesperson';
-import { SalespersonFormComponent } from 'src/app/shared/components/salesperson-form/salesperson-form.component';
-import { COMMON_WARNING_MESSAGE, COMMON_DISMISS_BUTTON_LABEL, COMMON_ERROR_MESSAGE } from 'src/text/messages';
-import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form/EntityFormDialogConfig';
-import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager.component-directive';
-import { ManagementSalespeopleService } from './management-salespeople.service';
 import { of } from 'rxjs';
+import { catchError, map, tap } from 'rxjs/operators';
+import { SalespersonFormComponent } from 'src/app/shared/components/salesperson-form/salesperson-form.component';
+import { Salesperson } from 'src/models/entities/Salesperson';
+import { COMMON_DISMISS_BUTTON_LABEL, COMMON_ERROR_MESSAGE } from 'src/text/messages';
+import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form/EntityFormDialogConfig';
+import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager/transactional-data-manager.component.directive';
+import { ManagementSalespeopleService } from './management-salespeople.service';
 
 @Component({
   selector: 'app-management-salespeople',
