+++ src/app/management/routes/shippers/management-shippers.component.ts
@@ -9,14 +9,14 @@ import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute } from '@angular/router';
-import { catchError, map, tap } from 'rxjs/operators';
-import { Shipper } from 'src/models/entities/Shipper';
-import { ShipperFormComponent } from 'src/app/shared/components/shipper-form/shipper-form.component';
-import { COMMON_WARNING_MESSAGE, COMMON_DISMISS_BUTTON_LABEL, COMMON_ERROR_MESSAGE } from 'src/text/messages';
-import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form/EntityFormDialogConfig';
-import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager.component-directive';
-import { ManagementShippersService } from './management-shippers.service';
 import { of } from 'rxjs';
+import { catchError, map, tap } from 'rxjs/operators';
+import { ShipperFormComponent } from 'src/app/shared/components/shipper-form/shipper-form.component';
+import { Shipper } from 'src/models/entities/Shipper';
+import { COMMON_DISMISS_BUTTON_LABEL, COMMON_ERROR_MESSAGE } from 'src/text/messages';
+import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form/EntityFormDialogConfig';
+import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager/transactional-data-manager.component.directive';
+import { ManagementShippersService } from './management-shippers.service';
 
 @Component({
   selector: 'app-management-shippers',
