+++ src/app/management/routes/users/management-users.component.ts
@@ -9,14 +9,14 @@ import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute } from '@angular/router';
-import { catchError, map, tap } from 'rxjs/operators';
-import { User } from 'src/models/entities/User';
-import { UserFormComponent } from 'src/app/shared/components/user-form/user-form.component';
-import { COMMON_WARNING_MESSAGE, COMMON_DISMISS_BUTTON_LABEL, COMMON_ERROR_MESSAGE } from 'src/text/messages';
-import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form/EntityFormDialogConfig';
-import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager.component-directive';
-import { ManagementUsersService } from './management-users.service';
 import { of } from 'rxjs';
+import { catchError, map, tap } from 'rxjs/operators';
+import { UserFormComponent } from 'src/app/shared/components/user-form/user-form.component';
+import { User } from 'src/models/entities/User';
+import { COMMON_DISMISS_BUTTON_LABEL, COMMON_ERROR_MESSAGE } from 'src/text/messages';
+import { EntityFormDialogConfig } from '../../../shared/dialogs/entity-form/EntityFormDialogConfig';
+import { TransactionalDataManagerComponentDirective } from '../../directives/transactional-data-manager/transactional-data-manager.component.directive';
+import { ManagementUsersService } from './management-users.service';
 
 @Component({
   selector: 'app-management-users',
