+++ src/app/management/routes/users/management-users.service.ts
@@ -10,7 +10,7 @@ import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-service-injection-
 import { ITransactionalEntityDataApiService } from 'src/app/api/transactional-entity.data-api.iservice';
 import { SharedDialogService } from 'src/app/shared/dialogs/shared-dialog.service';
 import { User } from 'src/models/entities/User';
-import { TransactionalDataManagerServiceDirective } from '../../directives/transactional-data-manager.service-directive';
+import { TransactionalDataManagerServiceDirective } from '../../directives/transactional-data-manager/transactional-data-manager.service.directive';
 
 @Injectable()
 export class ManagementUsersService
